Refactor list student
