+++ mobile/lib/features/student_management/presentation/widgets/student_item.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+import '../../domain/entities/student.dart';
+import '../pages/detail_student.dart';
+
+class StudentItem extends StatelessWidget {
+  final Student _student;
+
+  StudentItem(Student student) : this._student = student;
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      child: ListTile(
+        title: Text(_student.firstName.toString()),
+        subtitle: Text(_student.classroomId.toString()),
+        trailing: IconButton(
+          icon: Icon(Icons.arrow_forward),
+          onPressed: () {
+            Navigator.of(context).push(
+              MaterialPageRoute(
+                builder: (context) => StudentDetailPage(_student),
+              ),
+            );
+          },
+        ),
+      ),
+    );
+  }
+}
